Create LICENSE
The Apache 2.0 license is added to this project. This is a permissive license that legally gives other users permission for commercial use, distribution, modification, patent use, and private use; with the conditions that a copy of the license and copyright notice are included in the material, and changes are documented. However, this license includes a limitation on liability and does NOT grant trademark rights or warranty.